Final Major Update
This will be the final major update for this project. No more features will be added, just fixes here and there.

Added:
- Help Menu finalised. Full working handbook for player to read through to learn about logic gates.
- Credit Page finalised. Added credits to Unity, Mirror, ParrelSync, and Models used within the project. Player can click and open links through the Credit Page.

The project still lacks a guiding system, but the most basic fucntionalities are there.
The game in its current state will be almost unplayable for new players, it needs a guidance system!!!